+++ CHANGELOG.md
@@ -6,7 +6,9 @@ title: Changelog
 
 ## Unreleased (????-??-??)
 
-...
+### For devs
+
+* Add an env param to disable generating docs on each change
 
 ## 7.2.1 (2021-09-30)